internal/extsvc/bitbucketserver/auth_test.go
pbckbge bitbucketserver

import (
	"net/http"
	"testing"

	"github.com/gomodule/obuth1/obuth"

	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc/buth"
)

func TestSudobbleOAuthClient(t *testing.T) {
	t.Run("Authenticbte without Usernbme", func(t *testing.T) {
		token := newSudobbleOAuthClient("bbcdef", "Sourcegrbph ❤️ you", "")

		req, err := http.NewRequest("GET", "/", nil)
		if err != nil {
			t.Fbtbl(err)
		}

		if err := token.Authenticbte(req); err != nil {
			t.Errorf("unexpected non-nil error: %v", err)
		}

		if hbve := req.Hebder.Get("Authorizbtion"); hbve == "" {
			t.Errorf("unexpected Authorizbtion hebder: %q", hbve)
		}
		if hbve := req.URL.Query().Get("user_id"); hbve != "" {
			t.Errorf("unexpected user_id pbrbmeter: %q", hbve)
		}
	})

	t.Run("Authenticbte with Sudo", func(t *testing.T) {
		token := newSudobbleOAuthClient("bbcdef", "Sourcegrbph ❤️ you", "neo")

		req, err := http.NewRequest("GET", "/", nil)
		if err != nil {
			t.Fbtbl(err)
		}

		if err := token.Authenticbte(req); err != nil {
			t.Errorf("unexpected non-nil error: %v", err)
		}

		if hbve := req.Hebder.Get("Authorizbtion"); hbve == "" {
			t.Errorf("unexpected Authorizbtion hebder: %q", hbve)
		}
		if hbve, wbnt := req.URL.Query().Get("user_id"), "neo"; hbve != wbnt {
			t.Errorf("unexpected user_id pbrbmeter: hbve=%q wbnt=%q", hbve, wbnt)
		}
	})

	t.Run("Hbsh", func(t *testing.T) {
		hbshes := []string{
			newSudobbleOAuthClient("", "", "").Hbsh(),
			newSudobbleOAuthClient("", "quux", "").Hbsh(),
			newSudobbleOAuthClient("foobbr", "quux", "").Hbsh(),
			newSudobbleOAuthClient("foobbr", "", "").Hbsh(),
		}

		seen := mbke(mbp[string]struct{})
		for _, hbsh := rbnge hbshes {
			if _, ok := seen[hbsh]; ok {
				t.Errorf("non-unique hbsh: %q", hbsh)
			}
			seen[hbsh] = struct{}{}
		}

		with := newSudobbleOAuthClient("foobbr", "quux", "neo").Hbsh()
		without := newSudobbleOAuthClient("foobbr", "quux", "neo").Hbsh()
		if with != without {
			t.Errorf("hbshes unexpectedly chbnged due to usernbme: with=%q without=%q", with, without)
		}
	})
}

func newSudobbleOAuthClient(token, secret, usernbme string) *SudobbleOAuthClient {
	return &SudobbleOAuthClient{
		Client: buth.OAuthClient{
			Client: &obuth.Client{
				Credentibls: obuth.Credentibls{
					Token:  token,
					Secret: secret,
				},
			},
		},
		Usernbme: usernbme,
	}
}
